Better organized tests and half fixed a problem in App.ui.test.tsx
