+++ client/src/__tests__/integration-tests/App.integration.test.ts
@@ -1,145 +1,9 @@
-import App from '../App';
-import React from 'react';
-import { fireEvent, screen, waitFor } from '@testing-library/react';
-import '@testing-library/user-event';
-import userEvent from '@testing-library/user-event'
 import '@testing-library/jest-dom';
 import 'jest-extended';
 import nock from 'nock';
 import axios from 'axios';
 axios.defaults.adapter = require('axios/lib/adapters/http');
-import getFighters from '../lib/fetch-fighters';
-import { renderWithProviders } from '../util/test-utils';
-
-describe("Testing App.tsx UI/UX", () => {
-
-  describe("Testing Home page", () => {
-
-    it("Renders Home on page after loading", async () => {
-      renderWithProviders(<App />);
-      /**
-       * Cards on page aren't rendered until a database
-       * fetch is received.
-       *
-       * waitFor() pauses the test for up to 3000ms
-       * or until database fetch is received.
-       */
-      await (waitFor(() => screen.findByText(/Bayonetta/i), { timeout: 3000 }));
-      await screen.findByText(/Inkling/i);
-      await screen.findByTestId(/^joker-heart$/i);
-    });
-
-  })
-
-  describe("testing navbar links", () => {
-
-    it("Renders signIn component when Login link is clicked", async () => {
-      const user = userEvent.setup();
-      renderWithProviders(<App />);
-      const loginNav = screen.getByRole('link', { name: /^login$/i });
-      expect(loginNav).toHaveAttribute('href', '/registration/sign-in');
-      await user.click(loginNav);
-      await screen.findByTestId(/^sign-in-form$/i);
-    });
-
-    it("Renders Home component when Home nav is clicked", async () => {
-      renderWithProviders(<App />);
-      const homeNav = screen.getByRole('link', { name: /^home$/i });
-      expect(homeNav).toHaveAttribute('href', '/');
-      await userEvent.click(homeNav);
-      await screen.findByTestId(/^joker$/i);
-    });
-
-    it("Renders Favorites component when Favorites nav is clicked", async () => {
-      const user = userEvent.setup();
-      renderWithProviders(<App />);
-      const favoritesNav = screen.getByRole('link', { name: /^favorites$/i });
-      expect(favoritesNav).toHaveAttribute('href', '/favorites');
-      await user.click(favoritesNav);
-      await screen.findByText(/^Favorites Are Empty/i);
-      await screen.findByRole('link', {name: /^add some!$/i});
-    });
-
-    it("Renders signIn component when Login icon is clicked", async () => {
-      const user = userEvent.setup();
-      renderWithProviders(<App />);
-      await user.click(screen.getByTestId(/^profile-icon$/i));
-      await screen.findByTestId(/^sign-in-form$/i);
-    });
-
-    it("Renders Home when Smashmate title is clicked", async () => {
-      renderWithProviders(<App />);
-      const homeNav = screen.getByRole('link', { name: /^smashmate$/i });
-      expect(homeNav).toHaveAttribute('href', '/');
-      await userEvent.click(homeNav);
-      await screen.findByTestId(/^joker$/i);
-    });
-  });
-
-  describe("Testing favoriting/unfavoriting fighters", () => {
-
-    it("Favorites a fighter and they appear on Favorites page", async () => {
-      const user = userEvent.setup();
-      renderWithProviders(<App />);
-      await user.click(screen.getByText(/^home$/i));
-      const heart = await screen.findByTestId(/^pyra-heart$/i);
-      const favorites = await screen.findByText(/^favorites$/i);
-      await user.click(heart)
-      await user.click(favorites);
-      await screen.findByTestId(/^pyra$/i);
-    });
-
-    it("unfavorites a fighter and they disappear from the Favorites page", async () => {
-      const user = userEvent.setup();
-      renderWithProviders(<App />);
-      await user.click(screen.getByRole('link', { name: /^favorites$/i }));
-      const heart = await screen.findByTestId(/^pyra-heart$/i);
-      expect(heart).toBeInTheDocument();
-      await user.click(heart);
-      expect(heart).not.toBeInTheDocument();
-    });
-
-  });
-
-  describe("Testing fighterDetails", () => {
-
-    it("renders fighterDetails component when a fighter's card is clicked", async () => {
-      const user = userEvent.setup();
-      renderWithProviders(<App />);
-      await user.click(screen.getByText(/^home$/i));
-      const joker = await screen.findByTestId(/^joker$/i);
-      await user.click(joker);
-      const moves = await screen.findByText(/^Moves$/i);
-      const grabs = await screen.findByText(/^Grabs\/Throws$/i);
-      const dodges = await screen.findByText(/^Dodges\/Rolls$/i);
-      const stats = await screen.findByText(/^Stats$/i);
-      expect(moves).toBeInTheDocument();
-      expect(grabs).toBeInTheDocument();
-      expect(dodges).toBeInTheDocument();
-      expect(stats).toBeInTheDocument();
-    });
-
-    it("renders fighterDetails component when 'enter' is pressed on a focused card", async () => {
-      const user = userEvent.setup();
-      renderWithProviders(<App />);
-      await user.click(screen.getByText(/^home$/i));
-      const joker = await screen.findByTestId(/^joker$/i);
-      joker.focus();
-      expect(document.activeElement).toBe(joker);
-      await user.keyboard('abc');
-      expect(document.activeElement).toBe(joker);
-      await user.keyboard('{Enter}');
-      const moves = await screen.findByText(/^Moves$/i);
-      const grabs = await screen.findByText(/^Grabs\/Throws$/i);
-      const dodges = await screen.findByText(/^Dodges\/Rolls$/i);
-      const stats = await screen.findByText(/^Stats$/i);
-      expect(moves).toBeInTheDocument();
-      expect(grabs).toBeInTheDocument();
-      expect(dodges).toBeInTheDocument();
-      expect(stats).toBeInTheDocument();
-    });
-  })
-});
+import getFighters from '../../lib/fetch-fighters';
 
 describe("testing /api/get/fighters route", () => {
   afterEach(nock.cleanAll);
