+++ client/src/__tests__/integration-tests/components/registration/CreateAccount.integration.test.ts
