+++ client/src/__tests__/integration-tests/components/registration/SignIn.integration.test.tsx
